added error admin

+++ controllers/admin.js
@@ -6,7 +6,7 @@ exports.getAllProducts = async (req, res, next) => {
 
     res.status(200).json({ message: "Get all products", products: products });
   } catch (error) {
-    console.log(error);
+    return next(error);
   }
 };
 
@@ -16,9 +16,15 @@ exports.getProductById = async (req, res, next) => {
   try {
     const product = await Product.findByPk(productId);
 
+    if (!product) {
+      const error = new Error("Not found this product.");
+      error.code = 400;
+      throw error;
+    }
+
     res.status(200).json({ message: "Get product", product: product });
   } catch (error) {
-    console.log(error);
+    return next(error);
   }
 };
 
@@ -31,13 +37,14 @@ exports.postAddProduct = async (req, res, next) => {
       price,
       imageUrl,
       description,
+      userId: req.userId,
     });
 
     res
       .status(201)
       .json({ message: "Add product successfull", product: result });
   } catch (error) {
-    console.log(error);
+    return next(error);
   }
 };
 
@@ -47,6 +54,20 @@ exports.postEditProduct = async (req, res, next) => {
 
   try {
     const product = await Product.findByPk(productId);
+
+    if (!product) {
+      const error = new Error("Not found this product.");
+      error.code = 400;
+      throw error;
+    }
+
+    if (req.userId !== product.userId) {
+      const error = new Error(
+        "Not authenticated to delete product, please try again."
+      );
+      error.code = 401;
+      throw error;
+    }
 
     product.title = title;
     product.price = price;
@@ -59,13 +80,29 @@ exports.postEditProduct = async (req, res, next) => {
       .status(200)
       .json({ message: "Updated product successfull", product: result });
   } catch (error) {
-    console.log(error);
+    return next(error);
   }
 };
 
 exports.postDeleteProduct = async (req, res, next) => {
   const { productId } = req.params;
   try {
+    const product = await Product.findByPk(productId);
+
+    if (!product) {
+      const error = new Error("Not found this product.");
+      error.code = 400;
+      throw error;
+    }
+
+    if (req.userId !== product.userId) {
+      const error = new Error(
+        "Not authenticated to delete product, please try again."
+      );
+      error.code = 401;
+      throw error;
+    }
+
     const result = await Product.destroy({
       where: {
         id: productId,
@@ -76,6 +113,6 @@ exports.postDeleteProduct = async (req, res, next) => {
       .status(200)
       .json({ message: "Delete product successfull", product: result });
   } catch (error) {
-    console.log(error);
+    return next(error);
   }
 };
